+++ src/Messaging/RuntimeCaller/ZeroMQRuntimeCaller.php
@@ -44,6 +44,19 @@ class ZeroMQRuntimeCaller
     private $mapper;
 
     /**
+     * @param Param $param
+     * @return array
+     */
+    private function writeParam(Param $param)
+    {
+        return [
+            'n' => $param->getName(),
+            'v' => $param->getValue(),
+            't' => $param->getType(),
+        ];
+    }
+
+    /**
      * @param MessagePackSerializer $serializer
      * @param CompactTransportMapper $mapper
      * @param $socket
@@ -89,7 +102,7 @@ class ZeroMQRuntimeCaller
                         $target->getAction()
                     ],
                     'T' => $this->mapper->writeTransport($transport),
-                    'p' => $params,
+                    'p' => array_map([$this, 'writeParam'], $params),
                     'f' => $files,
                 ],
             ],
@@ -115,11 +128,10 @@ class ZeroMQRuntimeCaller
             $reply = $this->socket->recv();
             $response = $this->serializer->unserialize($reply);
 
-            if (isset($response['cr']['r']['E'])) {
-                throw new RuntimeCallException("Error response: {$response['cr']['r']['E']['m']}");
+            if (isset($response['E'])) {
+                throw new RuntimeCallException("Error response: {$response['E']['m']}");
             } else {
-                echo json_encode($response), "\n";
-                $return = $response['cr']['r']['R'];
+                $return = $response['cr']['r']['rv'];
                 $this->mapper->merge($transport, $response['cr']['r']['T']);
             }